Update source defined message format Move destination namespace names to i18n
